Split filter modal components
